Puts the helper methods into helpers.rb
